Add files via upload
Data from the bank's call center for March 1999. The data contains information about the time of call arrival at the call center (vru_entry), waiting time for a response to the call (q_time), service time (ser_time), the line to which the call was directed (vru+line), service type (type), the operator who answered the call (server), and other details.

It is necessary to estimate the distribution of service time (ser_time). This distribution may vary depending on the line to which the call was directed (vru+line), service type (type), and the operator who answered the call (server).